+++ src/renderer/src/components/editor/editor-panel-content-types.ts
@@ -1,0 +1,11 @@
+import type { GitDiffResult } from '../../../../shared/types'
+
+export type FileContent = {
+  content: string
+  isBinary: boolean
+  isImage?: boolean
+  mimeType?: string
+  loadError?: string
+}
+
+export type DiffContent = GitDiffResult
